Add Query component
